Don't cast angle to integer, so that it's possible to rotate less than one degree
BUG:156754

svn path=/trunk/koffice/; revision=876284

+++ krita/plugins/viewplugins/rotateimage/dlg_rotateimage.h
@@ -53,7 +53,7 @@ public:
     ~DlgRotateImage();
 
     void setAngle(quint32 w);
-    qint32 angle();
+    double angle();
 
     void setDirection(enumRotationDirection direction);
     enumRotationDirection direction();
